requirements.txt
antlr4-python3-runtime==4.7.2
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.2.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.1
cache-to-disk==0.0.3
certifi==2019.9.11
cffi==1.12.2
chardet==3.0.4
conda==4.7.12
conda-package-handling==1.6.0
cryptography==2.6.1
decorator==4.4.0
docutils==0.16
entrypoints==0.3
flake8==3.7.8
future==0.18.0
idna==2.8
importlib-metadata==0.23
ipython==7.8.0
ipython-genutils==0.2.0
isodate==0.6.0
isort==4.3.21
jedi==0.15.1
joblib==0.13.2
keyring==21.2.0
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.2
packaging==19.2
parso==0.5.1
-e git+git@github.com:tatsubori/pddl-lib.git@1d20563b9726f1d93686f69f7cc96a1575cebfbf#egg=pddlpy
pddls==0.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
pluggy==0.13.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
PySocks==1.6.8
pytest==5.2.1
python-jsonrpc-server==0.2.0
python-language-server==0.29.0
PyYAML==5.3.1
rdflib==4.2.2
readme-renderer==25.0
requests==2.21.0
requests-toolbelt==0.9.1
rope==0.14.0
ruamel-yaml==0.15.46
six==1.12.0
snowballstemmer==2.0.0
torch==1.3.1
tqdm==4.36.1
traitlets==4.3.3
twine==3.1.1
typed-ast==1.4.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.11.2
yapf==0.28.0
zipp==0.6.0
